Split files a bit, start working on entry api
